+++ HISTORY.rst
@@ -3,6 +3,12 @@
 History
 -------
 
+0.2.0 (2014-09-08)
+++++++++++++++++++
+
+* Revise `has_geolocation` method, remove property status
+* Added list filter class `GeocodedFilter`
+
 0.1.1 (2014-08-26)
 ++++++++++++++++++